LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/metadata/font.https.sub-expected.txt

1
2
3

PASS Same-Origin font
PASS Same-Site font
PASS Cross-Site font
PASS Same-Origin font: sec-fetch-dest
FAIL Same-Origin font: sec-fetch-mode assert_equals: expected "cors" but got "no-cors"
PASS Same-Origin font: sec-fetch-site
PASS Same-Origin font: sec-fetch-user
PASS Same-Site font: sec-fetch-dest
FAIL Same-Site font: sec-fetch-mode assert_equals: expected "cors" but got "no-cors"
PASS Same-Site font: sec-fetch-site
PASS Same-Site font: sec-fetch-user
PASS Cross-Site font: sec-fetch-dest
FAIL Cross-Site font: sec-fetch-mode assert_equals: expected "cors" but got "no-cors"
PASS Cross-Site font: sec-fetch-site
PASS Cross-Site font: sec-fetch-user
